refactor to create a kerchunk 'reader'
